backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==24.3
drf-yasg==1.21.7
requests==2.31.0
psycopg2-binary==2.9.9
Pillow==10.2.0
reportlab==4.0.7
python-dotenv==1.0.0
dj-database-url==3.0.1
